Replace 'killall -9' with 'killall' for non-Windows OSes
As per the current MongoDB docs and issue:
https://docs.mongodb.com/v3.4/tutorial/manage-mongodb-processes/#use-kill
https://github.com/mongodb-js/kill-mongodb/issues/6

Just not sure about possible downstream users of this, i.e. whether anyone using mongodb-runner for instance might be relying on it in some edge case. Perhaps a major semver bump?

EDIT: Probably not, as it's only a devDependency for now so probably not as big of a deal:
https://github.com/mongodb-js/runner/blob/414359f88d6f2f9d23badbbcaa6b2028f4445656/package.json#L58-L62